// src/types/report/trading.ts
export type TTradingBase = {
  storeId?: any;
  storeName?: string;
  totalOrderAtStore?: number;
  totalOrderTakeAway?: number;
  totalOrderDelivery?: number;
  finalAmountAtStore?: number;
  finalAmountTakeAway?: number;
  finalAmountDelivery?: number;
  totalBills?: number;
  totalSales?: number;
  totalDiscount?: number;
  totalSalesAfterDiscount?: number;
};